Move visualize into a separate package and remove features.
Removes the following features:
- Drawing of piecewise constant functions for Mesh2D
- Drawing of piecewise-linear DG functions for MeshTri

These should be implemented using InteriorBasis.refinterp instead.
